Hume-AI-Voice-Clone-Claude-Opus/requirements.txt
hume
pyaudio
requests
numpy
python-dotenv
